+++ Extra/Settings/Changelog.md
@@ -45,5 +45,9 @@ First released version!
 
 @ver v1.0.8
 ```diff
-+ Added color custumization to the map background, borders, and extra info section.
++ Added color custumization to the map background, borders, and extra info section. (thenose2003)
++ Fix bug with hover info sometimes showing on map when not hovering
++ Tenios map style
++ Toggle map setting
++ Fix scoreboard score setting
 ```